change double to long double and add gravity, coef friction and energi to tui_left

+++ includes/pendule.h
@@ -5,29 +5,29 @@ class Pendule {
     
 public:
     Pendule();
-    Pendule(int id, double m, double r, double theta0, double omega, Pendule* parent);
-    Pendule(int id, double m, double r, double theta0, double omega);
+    Pendule(int id, long double m, long double r, long double theta0, long double omega, Pendule* parent);
+    Pendule(int id, long double m, long double r, long double theta0, long double omega);
     ~Pendule();
 
     //Les gets
 
     int id();
-    double x();
-    double y();
-    double r();
-    double theta();
-    double oldomega();
-    double omega();
-    double m();
+    long double x();
+    long double y();
+    long double r();
+    long double theta();
+    long double oldomega();
+    long double omega();
+    long double m();
     Pendule* attacher();
     
     //Les set
     void id(int id);
-    void r(double r);
-    void theta(double theta);
-    void oldomega(double theta);
-    void omega(double omega);
-    void m(double m);
+    void r(long double r);
+    void theta(long double theta);
+    void oldomega(long double theta);
+    void omega(long double omega);
+    void m(long double m);
 
     void attacher(Pendule* attache);
     void print();
@@ -35,11 +35,11 @@ public:
 private:
 
     int identifieur = 0.;
-    double th = 0.;
-    double longeur = 0.;
-    double vit_angl = 0.;
-    double oldvit = 0.;
-    double masse = 0.;
+    long double th = 0.;
+    long double longeur = 0.;
+    long double vit_angl = 0.;
+    long double oldvit = 0.;
+    long double masse = 0.;
     Pendule* attache = NULL;
 };
